requirements/dev.txt
# requirements/dev.txt: development dependency requirements
-r common.txt

coverage==3.6
django-debug-toolbar==0.9.4
django-nose==1.2
flake8==2.0
pylint==1.0.0
tox==1.6.0

-r project/common.txt
-r project/dev.txt
